Add a Github action to run opa tests
Establising some CI for the rego policy files before we start adding
them. Also update the readme, and add some help to the Makefile.

(Maybe make is a little out of style, but I think it's useful as a
task runner for this kind of thing, and it helps avoid duplication
of logic in the GitHub action.)