Install opa under $HOME so we don't require sudo
It's motivated by the CI since we have to install opa every time,
but I wanted to make it work if used locally as well.

$(HOME)/.local/bin is in the path already in the ubuntu gitlab
runner. I'm assuming $(HOME)/bin is in your path locally, but I
don't care much since it's not hard to install opa wherever you
want.