Add make command to embed data into rego file
Also:
- Check in some generated data
- Add some example rego tests showing how we can use it

For testing purposes it will be useful to have realistic test data.
I don't think it's possible to access external data when running
`opa test` so this is a creative way to provide access a useful data
set that can be used to write tests against.

Possibly not the "normal" way that people do testing for opa, but I
think it can work well for us, so let's see how it goes.

The data checked in now comes from a nodejs builder pipeline run in
my local cluster.