Add Makefile script to help doing TDD using entr
The technique was described by @lcarva and I thought it was useful
enough to include here.

There some extra work to make it handle files being added and
removed. The -d option means entr will quit if it notices files
being added or removed and the loop forces the file list to be
refreshed.