Refactor to generalize
This moves the policy rules to `policies` package so we can generalize
over all policies. That is, in `main` we no longer need to call out each
policy by name.
The test for main is now a bit more fragile as changes to the policies
could fail the test, this is to assert that the output from main hasn't
changed in it's shape (say from a set to an array).