Run `opa check` in strict mode
Additional check we can make is to run `opa check`, this adds it as a
rule to the Makefile and refactors the GitHub workflow to run the `ci`
rule. There are a few[1] strict checks that are currently enforced.

The difference in running the `ci` rule is that instead of `test`
`test-quiet` is invoked.

[1] https://www.openpolicyagent.org/docs/latest/strict/