Add a check so CI fails if the docs need an update
The workflow is similar to the fmt check. The CI will fail if the
freshly generated docs don't match the version on disk.

(I considered comparing with HEAD so it disregards uncommitted
changes in index.md, but decide not to worry about it, since the
fmt check doesn't do that and it seems to not be a problem.)

There's some extra effort made so that the check does not modify the
generated doc file, similar to the fmt check, even though it
probably wouldn't matter much if it did.