Support "SKIPPED" test result status
With this we support the "SKIPPED" test result status. The definition of
`deny` for test results is changed to not being a "FAILURE" or "ERROR",
instead of being "SUCCESS"; and a new `warn` is added for tests with the
status of "SKIPPED".

Also denies any tests resulting in an unsupported result.

Ref. https://issues.redhat.com/browse/HACBS-740