Validate required tasks present in attestation
This adds two policy checks: `tasks_missing` if there are no tasks in
the PipelineRun attestation; and `tasks_required` if the required tasks
are not present in the PipelineRun attestation.

In addition to custom annotation for `rule_data` in the rule scope, this
also allows custom `rule_data` annotation at the package scope that is
nested within the rule's short name. In this case this was necessary so
that the list of required tasks is not repeated between the test and the
rule implementation. This allows us to define `rule_data` for the
`tasks_required` rule and the tests in the single location.

Ref. https://issues.redhat.com/browse/HACBS-676