Order OPA inspect data within doc template
Seems that the order of keys in data from opa inspect can vary, it is
also unclear if unmarshalling the resulting JSON can lead to it being
reordered due to golang's map hashing algorithm. To achieve
deterministic output in the docs this sorts the data by location: rule
file path and line number of the rule's definition; in the documentation
template.