Render documentation from local source always
To prevent checking out the repository twice we can build from local
checkout in all cases (for `docs-render` and `docs-preview` targets).
This both simplifies the rendering and makes it deterministic, as we
didn't specify `--fetch` which would go through the Antora cache and
possibly use an older version of the documentation.

This way we also don't require the GitHub Action to render, and can use
the `docs-render` target from the Makefile, so one less item to
maintain.