Sort the annotations json so it stays the same
For some reason opa inspect outputs the annotations in random order
which means we get useless diffs every time it's refreshed. Make it
deterministic by sorting the annotations list.

Added the jq -S option as well to sort the keys. I'm not sure that
is needed but it can't do any harm afaict.