Rework rule include/exclude handling
There's a new mechanism for selecting which rules are included and
which are skipped.

You can specify which rules should run in more flexible ways, e.g.
by naming the rules you want, or the rules you want to skip, or a
combination of both. It's supposed to work reasonably as you would
expect, but note that exclusions have precedence over inclusions.

Two new policy configuration values are introduced:
* data.config.policy.include_rules
* data.config.policy.exclude_rules

include_rules and exclude_rules are lists of strings similar to the
existing non_blocking_checks list. The non_blocking_checks list
still works for backwards compatibility reasons but we should
consider it deprecated and remove it soon once the ECP and the
ec-cli are ready use the new fields.

You can now include or skip individual rules (not just packages).

There's a wildcard "*" which matches any package, so you can use it
to express "exclude all" or "include all".

Additionally there's another policy configuration value:
* data.config.policy.collection

If this is present then the include_rules and exclude_rules are set
to one of the options predefined in data/rule_collections.yml. This
allows certain combinations of rules to be given a name and
a description.