Add a new way to include/exclude rules
There are two new configuration values:
- data.config.policy.include_rules
- data.config.policy.exclude_rules

These are lists of strings similar to the
existing data.config.policy.non_blocking_checks, and actually
exclude_rules is quite similar to non_blocking_checks in function.

One thing that is new is that you can specify a '*' which acts as a
wildcard to indicate it should match every package. (Using that in
the exclude_rules would skip every rule which is probably not a
useful thing.)

This is somewhat experimental at this stage. Note that it currently
doesn't have the individual test names in the test package the way
that non_blocking_checks does.