Include abspath in antora fileSrc
This avoids a error when antora wants to report a problem, somewhere
in the reshapeFileForLog function in node_modules/@antora/logger/lib/logger.js

(Side note: While debugging this I realized it looks like we're
setting the "origin" key incorrectly. It's supposed to be an object
with a bunch of keys in it rather than a string. Maybe that doesn't
matter though so I'm not doing anything about it here.)