Include abspath in antora fileSrc
This avoids a error when antora wants to report a problem, somewhere
in the reshapeFileForLog function in node_modules/@antora/logger/lib/logger.js

Ps, while debugging this I realized it looks like we're setting the
"origin" key incorrectly. It's supposed to be an object with a
bunch of keys in it rather than a string. Maybe it doesn't matter
though so I'm not going to do anything about it here.