Required tasks policy with invocation parameters
The `sanity-label-check` Task can be parameterized with
`POLICY_NAMESPACE` parameter, which can have values `required_checks`
and `optional_checks`. When the task is run in the pipeline it is given
then name `sanity-label-check` and `sanity-optional-label-check`. Note
this is the given name for a `Task` within the `PipelineRun`. We assert
the name of the referenced task from the Task Bundle, which in both
cases is `sanity-label-check`.

So to distinguish between this two cases in addition to requiring a task
by name, this allows to require a task by a parameter and value.

So the required task list now can include:
 - `sanity-label-check[POLICY_NAMESPACE=required_checks]`
 - `sanity-label-check[POLICY_NAMESPACE=optional_checks]`

This resolves the first part of HACBS-1166.

Ref. https://issues.redhat.com/browse/HACBS-1166