Check provenance of test and java TaskRun results
With this the `TaskRun` results considered by the `test` and `java`
policy rules need to be from a `TaskRun` loaded from via a Tekton
Bundle, i.e. not defined on the cluster or inline in a `PipelineRun`,
and the image reference from which the `Task` was loaded needs to be one
from the list of accepted bundles (`acceptable_tekton_bundles.yml`).

The caveat with the `java` policy rules is that there is no policy
denial resulting from omission of the result. This is relevant here
because the omission of the result could be due to the `TaskRun` result
resulting from an inline, cluster or unacceptable bundle.

Ref. https://issues.redhat.com/browse/HACBS-995