Tolerate missing or different tag in subject
When comparing the subjects given `IMAGE_URL` and `IMAGE_DIGEST` task
results with the attestation subject in cases when `IMAGE_URL` contains
the image tag, and the attestation subject.name doesn't we still don't
want to report a violation if `IMAGE_DIGEST` matches subject.digest.

Given that we're now removing strings from subject and results to
compare without tags, we also want to be careful not to compare empty
strings and consider them valid. I think.