Remove support for multi-stream bundle repos
The task Tekton bundles produced by the
redhat-appstudio/build-definitions repository used to have a very unique
structure. All the task definitions were split into chunks of 10, a
bundle would then created for each chunk, finally all bundles would be
pushed to the *same* OCI repository. To differentiate between the
different chunks, a digit suffix was added to the tag of each bundle.
So for the commit "abc", three Tekton bundles would be pushed to the
same repo, "abc-1", "abc-2", and "abc-3".

In order to properly determine if an acceptable bundle was being used,
the policies had to understand this structure. This was done by
dynamically splitting the list of bundles by the tag pattern, so "abc-1"
and "bcd-1" were on the same sub-list, but "abc-2" and "bcd-2" were on a
separate sub-list. Each sub-list was dubbed a "stream". This approach
was good enough in most cases. Its main drawback was the fact that it
was very specific to the build-definitions repository and unlikely to
ever be adopted by any other repository.

The build-definitions repository has moved away from this structure
(great!). Now each bundle contains a single task, and, more importantly,
each bundle is pushed to its own repository.

Since this logic is now irrelevant, it is with great pleasure removed
from the code based due to its complexity. (Even explaining the feature
itself is difficult!)

https://issues.redhat.com/browse/HACBS-1364

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>