Workaround conftest push error
I've observed that doing a conftest push of the exactly same bundle
(i.e. with an identical digest) and a different tag causes an error
like this:

  Error: push bundle: pushing manifest: failed to read expected number
  of bytes: unexpected EOF

I'm not sure if this is a bug in quay.io or in conftest or a
combination of both, but I'd like to avoid it with this workaround
that adds an extra data file at build time to ensure each build
produces a unique digest.

As a bonus, it exposes the git sha and a build timestamp to rego,
which could perhaps be used in interesting ways.

If the root cause of the bug is discovered, the workaround could be
removed.

I'm not sure if we will hit this problem already with the policy
bundle pushes. If we do then the same kind of trick could be used
there too maybe.