Smarter bundle pushing
Move the bundle pushing to a bash script that detects if a push is
needed. Copy the technique used in build-definitions repo, see
hack/build-and-push.sh.

Also change how the hack/pr-infra-deployments.sh is called. Call it
from the update-bundles script instead of from its own workflow step
so it's easier to create the PR only when it's needed.

I've included some inconsequential changes in the data and rego so
fresh bundles could be created when this gets merged.