Execute `opa test` once for coverage report
We report coverage in two ways by outputting an error if code coverage
of 100% is not met to the stdout and by generating the `coverage.json`
file. Now this requires executing `opa test` twice.

This removes the `coverage-report` target and introduces the generation
of `coverage.json` in the `quiet-test` target, removing the need to run
`opa test` twice for the coverage report.