Force bundle push with fresh bundle digest
It's a long story, but this is a hacky but effective way to ensure a
bundle push will be triggered with image digests that are new and
unique. If we ever need to do it again we can bump the number.

(Previously I've added some random comment or whitespace change, but
I didn't want to keep doing that.)

Related to https://issues.redhat.com/browse/HACBS-1603