Create a new "minimal" collection
The idea is that the policy rules in this collection should be easy
for the "five clicks and five minutes" user to understand and abide
by. It's likely to become the default config for brand new users.

I haven't a lot of time considering what should be included. I
figure it can be refined in later commits. However:

- Acceptable bundle checks are out, but task and base images
  prefixes checks are in
- SLSA checks about git repos are in since it seems like they'll
  be easy to pass
- All testing is currently out, mainly because I'm not sure if I
  can use an "include" in the config for individual test names.

Note that the stated requirement in Jira is to include a CVE scan,
so this will need some followup work later once we do have a way to
specify the inclusion of individual tests by name.

JIRA: HACBS-1615