Add operator-framework/opm to allowed registries
This is needed in particular for FBC builds. As a short term
solution let's just add it to the default allowed list.

In the longer term there should be some thinking about what useful
defaults would be and how we might maintain different versions of
these lists, e.g. for RH internal use and for "5 clicks" users.