Mention the built-in defaults in rule data yaml
Make it a little easier for users to know there are some other rule
data keys that they could override if they want to.

The initial motivation was to make it easier for users to be aware
of the two rule data keys related to the cve policy, since
customizing those is a thing that users might want to do. This is
documented in the rule itself, but I figure there's no harm in
making it discoverable in other ways.

Also set the default value of warn_cve_security_levels explicitly to
an empty list. It makes no functional difference, but it makes it
easier for someone reading the code to understand what is going on.

(This does bring up the question of why some keys have default
values in rule_data.rego and others don't, but let's consider that
later.)