Add default cve config to rule_data file
Show what the defaults are but leave them commented out. The idea is
to make it easy for someone to modify the file to suit their
preference, but avoid the possibility that the config values might
obscure the real defaults.

The values shown should match what is defined in
policy/lib/rule_data.rego.