Add sanity check for infra-deployments update
With this the manifests of the image referenced by tag and the image by
digest are compared to make sure they point to the same image.

We happened to have, for currently unknown, reason a change in
infra-deployments that had mismatched tag and digests. This was not
caught by e2e tests (at the time), but subsequently caused tests
failures. The digest in question does not correspond to any known
digests and the tag, derived from the manifest's label is known and
correct.