Revise rule titles and short names
The idea is that the title and name should describe what is being
checked, rather than describe the failure condition. This means the
title text makes more sense when used in the UI if the rule is
passing or if the rule is failing.

I'm also changing the short name for consistency, even though the
short name is not as exposed via the web UI.