Update description to not require a link
Some policy rules include asciidoc links in their description. The docs
building process replaces them with the correct link markup. However,
the description is used in other contexts which are agnostic to our
building process, e.g. `ec validate image --info`.

In https://issues.redhat.com/browse/HACBS-2102, a change to the ec-cli
is being made to remove such asciidoc markup and leave only the text
behind.

This change updates the wording of the description to still make sense
if a given part of the text is not a link.

This also removes some outdated references to acceptable bundles. (Not
all policy rules that check tasks require them to come from acceptable
bundles.)

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>