Don't push already pushed tags
The `https://quay.io/api/v1/repository/hacbs-contract/tag?specificTag`
API call requires authentication that we do not provide. This results in
valid JSON and `curl` exiting with `0` as the `401` HTTP status does not
result in non zero exit code with `curl`.

The `jq` following it will not find the tag in the JSON as it parses the
JSON error response and the bundle will be pushed regardless of the tag
existence.

This, for unknown reason, results in a different manifest digest
overwriting the previous tag.

To fix this, this uses `skopeo list-tags` instead, which will use the
existing Docker config credentials as needed.