Allow more than one buildah task
Given that some projects might include the buildah task more than once,
this opts to allow it and enforce the existing rules on all of them.

The message for missing parameters to the buildah task was modified to
include all the tasks that miss their parameters.

The message for using non-local Dockerfile was not changed.

Ref. https://issues.redhat.com/browse/HACBS-2280