Disable pipeline definition verification
A recent change has caused the `tkn` cli version to be used during the
validate-acceptable-bundles.sh script to be very recent. This is great.
However, it uncovered an issue with that script. The previous version
used was v0.18.0. In that version a kubeconfig was required to run the
`tkn bundle list` command. This, in turn, caused the script to always
succeed! The call is followed by `|| true` and stderr is muted.

This commit disables that check for now to avoid missing updates from
build-definitions. It'll be re-enabled as part of:
https://issues.redhat.com/browse/HACBS-2391

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>