Fixup latest digest for sast tasks
The sast-go and sast-java-sec-check tasks had a change that caused their
definition to be equal to one of its previous versions. Because of this
`ec track bundle` did not re-add entries for those bundles, as it is
expected. The issue is that in this case, this behavior doesn't make
sense as we want the old digest to be treated as the most up-to-date
task definition.

`ec track bundle` should be smarter about this. This will be done in:
https://issues.redhat.com/browse/HACBS-2394

This commit manually changes the affected entries.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>