Check OLM bundle image's manifests has pinned refs
Given the content of the OLM bundle image manifest files in input and
the presence of the `operators.operatorframework.io.bundle.manifests.v1`
label on the image this checks if all the references mentioned in the
Python checker[1] are pinned, i.e. they contain the digest.

Ref. https://issues.redhat.com/browse/HACBS-2379

[1] https://github.com/containerbuildsystem/operator-manifest#pull-specifications