Add policy rule to check deprecated att format
In HACBS-2390, EC CLI is modified to provide attestations in a different
format. In order to warn users in advance, a new policy rule checks if
the old format is still being used. The policy rule changes from a
warning to a violation on 2023-09-31, allowing users to adopt the new
format in about two months. By then, it is expected that all policy
rules, including the ones defined in this repo, have been updated to
support the new format.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>