SLSA source correlation check
Given the attested source code reference and the source image code
reference this check will make sure of their equality.

A warning is issued for source image data not including the source
reference.

A failure will be issued if no source reference is present in the
materials section or if the materials does not match the expected image
source reference for SLSA Provenance v0.2.

Similarly, a failure will be issued if no source reference is present in
the resolvedDependencies or if the resolvedDependencies does not match
the expected image source reference for SLSA Provenance v1.0.

Ref. https://issues.redhat.com/browse/HACBS-2247