Tolerate missing or superfluous `.git` suffix
At least in the e2e tests, and recent issue in Chains that was fixed
point to issues when the expected source, i.e. the one passed on via
ApplicationSnapshot, or the attested source, i.e. one in the materials
or resolvedDependencies, do no contain the `.git` suffix or have
superfluous `.git` suffix in the repository URL.
This allows for some leniency in the repository URL.