Mention the 'unknown' CVE severity in descriptions
Actually there's no functional change needed for EC to support the
new CVE severity, but we mention in the documentation and commentary
what the expected value for CVE severity are, so this patch updates
that.

Additionally I added 'unknown' to the dummy data in the tests, and
added some test coverage so it includes some output for the
'unknown' results. It's also not really needed, but I think it helps
communicate the intention for future readers of the code.

Resolves: https://issues.redhat.com/browse/EC-36