slsav1.0 refactor (#733)
Refactor all policies that verify an attestation to work with slsav1.0.

IMPORTANT: The task information here is pulled from the content field in resolvedDependencies. It is the taskRun that is stored there.
This mostly changes libraries to support accessing v0.2 and v1.0 attestations and updating the policy tests to support v1.0.
Update regal lint to warn on line-length too long. opa fmt in many cases is ok with a line-length that regal lint is not. So there is a conflict between the two.

https://issues.redhat.com/browse/HACBS-2303