Handle partial input.image.source data
The code assumed if `input.image.source` was present that the data
structure would always contain `url` and `revision` properties. Those
might be omitted as they're user-controlled inputs, e.g. from the
ApplicationSnapshot.
If those properties were omitted then the helper function
`_expected_sources` would return empty set as it would try to operate on
undefined data, and iterating over an empty set would lead to no checks
being performed.
This uses `object.get` instead of direct access to return a default
value if the properties are missing, so that the set should be populated
and the checks should fail to match.

ref: https://issues.redhat.com/browse/RHTAPBUGS-883