Ensure unknown tasks fail acceptable bundle checks
The attestation_task_bundle.task_ref_bundles_acceptable check should
produce a failure if the task bundle reference is not found at all
in the available `data.task-bundles` data. Previously it was only
failing if the repo was present in data.task-bundles and the task's
digest wasn't in the list.

Also some extra test coverage is added to exercise the different
failure conditions.

Resolves: https://issues.redhat.com/browse/EC-206