Add support for informative tests
Some tests are required to be executed, but their results do not have to
be successful. They are meant to inform the user that their results need
some attention, but they are non-blocking. An example of such test is a
certain code scanner which may produce a high number of false positives.
These tests are referred to as informative.

This commit modifies the `test` package to more easily support these
types of tests.

Also, since `test.no_skipped_tests` has been modified to emit a
violation instead of a warning, an effective date to a month from now
has been added. This should allow users to gradually address any issues
before they become blocking.

Ref: EC-224

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>