Do not rely on build task label for FBC images
When computing the key to select the set of required tasks we look at
the built task's or the pipeline label. Given that the build task for
the FBC and the Docker builds are the same, i.e. have the same label. We
cannot rely on the task's label, instead we this uses the image label
`operators.operatorframework.io.index.configs.v1` to classify a
attestation as being an attestation of a FBC image, and subsequently
select the set of required tasks for the `fbc` pipeline.

reference: EC-85