Use oras to build OPA bundles
The previous approach of using `conftest push` does not work now that EC
provides custom rego functions. We could wrap the `conftest` command in
EC CLI, like we did for OPA via `ec opa`. However, that is a bit heavy
handed.

The OPA docs[1] suggest using `opa bundle` to generate a tarball with
the compiled policy rules. Then, use `oras push` to upload the tarball
to an OCI registry.

The problem with the approach above is that when the OCI image
containing the policy rules is fetched, the output is a tarball instead
of the expected rego files. This causes conftest to not see the policy
rules. I considered adding support in the EC CLI to handle this use case
by simply expanding any tarballs that may have been fetched. But... this
means older version of the EC CLI won't be able to consume the newer
policy bundles.

Instead, this commit opts to use `oras push` directly on the rego files.
This creates a policy bundle that works as expected by the EC CLI.

There is benefit in running the `opa build` command since it compiles
the policy rules and verifies their integrity. For example, it checks if
the policy rules use an unknown function. This is useful. So, as a
sanity check, we now run `ec opa bundle` to ensure integrity, but the
output is discarded.

[1] https://www.openpolicyagent.org/docs/latest/management-bundles/#building-and-publishing-policy-containers

Fixes #817

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>