Handle duplicate acceptable bundle records
Currently the acceptable bundles can hold for the same image repository
multiple records with the same digest, they would also have the same tag
but their effective_on would differ. This would cause the rule to report
that the task bundle used is out of date as it would find two records
one of which could be newer than the other.

reference: EC-316