Add policy rule to enforce Task ref pinning
Ref: EC-257

This change adds a new attribute to the result returned by the
`task_ref` helper function, `pinned`. This indicates whether or not the
an immuntable reference is used.

For example, a git commit ID when using the git resolver is considered a
pinned reference, while a branch name is not.

A previous approach was to introduce a new helper method, e.g.
`is_pinned`. However, the a couple of things became obvious:

* The `task_ref` helper function already has the needed context, e.g.
  it knows that it's dealing with a git reference, so it can check if
  that's a git commit ID or a branch name.
* Being pinned is a property of the Task reference. It makes sense for
  the representation of the Task reference (what `task_ref` returns) to
  include this information.

As part of this, I also re-structured the unit tests in
`refs_test.rego`. Hopefully, this makes them easier to maintain.

This commit came out larger than I expected, but I can't think of a
reasonable way to split it up further.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>