Add initial acceptance tests
These tests run most of the policy rules against a pre-recorded
policy-input from the golden-container image.

The test utilities are very minimal. Just enough was implement to
executed a single test scenario.

It is expected that the `acceptance` module will grow over time and will
have a more formal structure.

Ref: EC-333

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>