Consolidate pruning of stale acceptable bundles
The list of acceptable bundles is time based. The presence of a record
is not a sufficient indicator for it being valid. As time goes on and
the effective_on date of records is reached, older entries fall off the
list. Using those records is akin to using a record that is not on the
list at all.

This commit introduces a helper function to prune old records from the
list before the list is consumed. This consolidates the decision of
which records should be ignored in a single place.

This fixes the related bug by ensuring stale records are used
consistently across the code base.

It also removes/modifies certain tests which were asserting incorrect
behavior.

Ref: RHTAPBUGS-1099

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>