Trusted Artifact chain policy
Given the change in[1], and the new data available in the attestation,
we can create a chain of PipelineRun tasks that contributed to the image
artifact being built.

Namely, the `*_ARTIFACT` parameters of a task should reference the
`*_ARTIFACT` results of a task whose artifact(s) it depends on.

Given such a chain, we can examine if all the Tasks comprising that
chain are acceptable.

This assumes that any acceptable Task conforms to using _only_ the
trusted artifacts when performing the its undertaking.

reference: https://issues.redhat.com/browse/EC-252

[1] https://github.com/redhat-appstudio/build-definitions/pull/712