wip: Set effective_on dynamically
This is a proof of concept for an idea discussed in the community
meeting. It leverages existing functionality in the EC CLI to demote
certain violations to warnings. So far, this has been used by setting
a custom annotation, `effective_on`, on the policy rule metadata. The
function `lib.result_helper*` takes this value and adds it to the
violation object.

In this commit, the policy rule dynamically determines what the
effective_on value should be based on a given data item, instead of
grabbing it from the policy rule's metadata.

This effectively means we can have time-based policy data items. This is
a simplistic example. I plan on showcasing more advanced ones shortly.

UPDATE: Also added an example of how this can be used in acceptable
bundles.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>