Labels release rules support time-based data
With this commit, the policy rules in the `labels` release package now
support time-based violations. Each item in the list of labels from the
data source can include an `effective_on` attribute. If so, the value is
added to the result of violations/warnings. The caller, i.e. EC CLI, is
responsible for processing this value and potentially downgrading
violations to warnings.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>