Fix update-bundles.sh script
The conditional branch to skip re-pushing a bundle required the
`ENSURE_UNIQUE` environment variable to be set to `"1"`. A quick find
across the EC repos didn't find any occurrences of this env var being
set, at all.

The `ENSURE_UNIQUE` seems like the remnants of a previous iteration of
this script that is not longer used.

This commit removes support for `ENSURE_UNIQUE`. It also moves registry
query to its own line instead of inlininig it in the if-statement for
readability purposes.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>