Add source_image package
This commit adds a new policy package, source_image. It leverages the
newly added `ec.oci.image_manifest` rego function to verify the source
image reference found in the SLSA Provenance is valid.

Because it relies on a newly added custom rego function, the package is
added to the `beta` namespace. At some point in the future, it will be
moved to the `release` namespace. The benefit of adding this package now
is that others can more easily verify its behavior.

Ref: EC-235

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>