Fix test with json patch
Prior to this commit, the `manifest` object was being patched twice. The
first time, it replaced the annotations with one value. The second time,
it replace the annotations with another value. In other words, the first
operation was completely undone by the second operation. This was not
the intention of the test.

This commit changes the behavior to replace specific attributes on the
annotations object.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>