Document trusted Tasks list
Documents to the new concept of trusted Tasks, and the trusted Task list
format, while noting that previous concept and data format for
acceptable bundles is kept only for backward compatibility.

The list of acceptable bundles is no longer published as it relies on
the example data. Keeping it on the website could be misleading as it is
not up to date.

Files were renamed and moved so that the acceptable bundles is not in
their file name.

Reference: https://issues.redhat.com/browse/EC-389