Fetch SBOM from OCI blob
This is using `ec.oci.blob` function to fetch the SBOM from an OCI blob,
expecting a JSON string to be present, which will be unmarshaled and
stored in the `data.lib.sbom.cyclonedx_sboms` array.

This all will happen only if the `beta` source directory is included in
the policy sources.

Reference: EC-499