Prevent release with quay expires-after set
The quay.expires-after label is used by quay.io to indicate the
image doesn't need to be kept indefinitely. The goal is to prevent
unintentionally releasing images with that label is set.

Uses the newly introduced `pipeline_intention` rule data to
determine if we're running the check in a release pipeline.

Ref: https://issues.redhat.com/browse/KFLUXBUGS-1223