Ensure schedule checks apply only to releases
Make it so the "no Friday deploys" rule (for example) is used only
for EC checks in release pipelines, and not for EC checks in regular
pre-merge or post-merge CI pipelines.

Whether we're in a release pipeline is determined by checking the
value of the new `pipeline_intention` rule data key.

See also:
* https://github.com/enterprise-contract/ec-cli/pull/1604
* https://github.com/konflux-ci/release-service-catalog/pull/409

Notes:
* Set a null default for the pipeline_intention rule data value
  since the standard default of an empty list seems less useful.
* This should allow the revert here to be unreverted:
  https://github.com/release-engineering/rhtap-ec-policy/pull/12

Ref: https://issues.redhat.com/browse/EC-293