Use Task name in term for trusted_task.trusted
When a policy rule emits a violation/warning that contains a `term`, the
`term` can be used in an exclude/include clause in a policy config.
This is meant to provide fine-grained control over certain policy rules,
e.g. ignore a violation but just for a particular Task.

Given how it's used, a `term` should be a piece of information that is
not possible to be changed without detection. Otherwise, it becomes
trivial to silently apply a certain exclusion to an unintended target.

A Task name is a good candidate for a `term` value. To change the Task
name, a change to the Task definition is required. Given our concept of
trusted Tasks, this can be easily detected.

A PipelineTask name, on the other hand, is completely within the control
of the Pipeline author. Since we do not establish any trust on Pipeline
definitions themselves, a Pipeline author can change the PipelineTask
name without raising any issues in EC. (This is as designed because we
take the approach of inspecting the contents of the Pipeline rather than
trusting specific implementations of the Pipeline.)

For this reason, the PipelineTask name should naver be used as a `term`
value.

This commit changes the `trusted_task.trusted` policy rule to use the
Task name as the `term` instead of the PipelineTask name. To avoid
confusion, the violation message is also updated to include both the
Task and PipelineTask name.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>