Add trusted.trusted_parameters policy rule
This policy rule is responsible for identifying when a trusted builder
Task uses the result from an untrusted Task.

The parameters inspected are scoped to values that look like a digest.
This is because, at least with the SLSA Provenance 0.2 format, we cannot
easily tell where a paramter came from (Pipeline parameter, default Task
Parameter, hard-coded, from another Task, etc). The new policy rule
makes an assumption that if a parameter includes a digest, that digest
must be accounted for within the Pipeline.

Fixes #1012
Ref: EC-658

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>