Edge case with Tasks that have same effective_on
If two versions of the Task have the same effective_on one of them will
be considered out of date, this is a bit confusing as the latest version
could be the one that is deemed out of date.

This changes so that if the proposed out of date task record and the
examined task record have the same exact effective_on the task is not
considered out of date.

We have this situation now, as the two top records for the buildah task
have the same effective on:

```yaml
  oci://quay.io/redhat-appstudio-tekton-catalog/task-buildah:0.1:
    - effective_on: "2024-06-29T00:00:00Z"
      ref: sha256:b8d50716c01160be5deb27dd30f79ae33cc89c72952285cd850323fa0a88b8db
    - effective_on: "2024-06-29T00:00:00Z"
      expires_on: "2024-06-29T00:00:00Z"
      ref: sha256:9f698f0903234093711ab2310891239fdcd6cfda5407f881822e6fa2cb05f771
```

and the latest digest `b8d5071` is ruled to be out of date.